Remove Lamp and add direct Light/Lights getters
It's confusing to have two different light types, especially when one of
them provides very little value. I've removed Lamp and the getter for it
on HueClient. In its place I've added a getter for the Light type, which
makes it far more discoverable than having to retrieve the bridge and then
access lights. A getter for individual lights was also added.