Fixed refresh token not used when token expires
When the access token expires after about 7 days, the refresh token, which expires in about 112 days, should be used to request a new access token. However, existing logic only evaluates to true when the refresh token expires. As such, after 7 days, an exception will be thrown stating that both tokens have expired. This commit fixes this logic to work as intended.